Fix parser register

+++ worker/worker.go
@@ -45,7 +45,7 @@ func (w Worker) Start() {
 	logging.WorkerCount.Inc()
 
 	for i := range w.queue {
-		if err := w.process(uint64(i)); err != nil {
+		if err := w.process(i); err != nil {
 			// re-enqueue any failed job
 			// TODO: Implement exponential backoff or max retries for a block slot.
 			go func() {
